refactor(room-node): extract shared room runtime
