+++ components/esp-openclaw-room-node/capture_mem.h
