+++ components/esp-openclaw-room-node/capture_thread.h
