+++ components/esp-openclaw-room-node/capture_utils.h
