+++ components/esp-openclaw-room-node/room_canvas_node_cmd.c
@@ -6,6 +6,7 @@
 #include "cJSON.h"
 #include "esp_check.h"
 #include "room_canvas.h"
+#include "room_canvas_internal.h"
 
 #define TAG "room_canvas_cmd"
 
@@ -32,9 +33,25 @@ static cJSON *parse_params(
     return params;
 }
 
+static bool only_fields(cJSON *params, const char *const *allowed, size_t count)
+{
+    for (cJSON *field = params->child; field != NULL; field = field->next) {
+        bool found = false;
+        for (size_t i = 0; i < count; ++i) {
+            if (field->string != NULL && strcmp(field->string, allowed[i]) == 0) {
+                found = true;
+                break;
+            }
+        }
+        if (!found) return false;
+    }
+    return true;
+}
+
 static esp_err_t handle_present(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -45,6 +62,11 @@ static esp_err_t handle_present(
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    static const char *const fields[] = {"url", "placement"};
+    if (!only_fields(params, fields, 2)) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.present contains an unknown field");
     }
 
     cJSON *url = cJSON_GetObjectItemCaseSensitive(params, "url");
@@ -62,6 +84,10 @@ static esp_err_t handle_present(
         url != NULL ? url->valuestring : NULL,
         out_payload_json,
         out_error);
+    if (err == ESP_OK) {
+        room_canvas_bind_owner_session(
+            room_canvas_surface_id != NULL ? invocation->session_key : NULL);
+    }
     cJSON_Delete(params);
     return err;
 }
@@ -69,6 +95,7 @@ static esp_err_t handle_present(
 static esp_err_t handle_navigate(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -79,6 +106,11 @@ static esp_err_t handle_navigate(
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    static const char *const fields[] = {"url"};
+    if (!only_fields(params, fields, 1)) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.navigate contains an unknown field");
     }
     cJSON *url = cJSON_GetObjectItemCaseSensitive(params, "url");
     if (!cJSON_IsString(url) || url->valuestring == NULL ||
@@ -88,6 +120,10 @@ static esp_err_t handle_navigate(
     }
 
     esp_err_t err = room_canvas_present(url->valuestring, out_payload_json, out_error);
+    if (err == ESP_OK) {
+        room_canvas_bind_owner_session(
+            room_canvas_surface_id != NULL ? invocation->session_key : NULL);
+    }
     cJSON_Delete(params);
     return err;
 }
@@ -95,6 +131,7 @@ static esp_err_t handle_navigate(
 static esp_err_t handle_hide(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -102,9 +139,14 @@ static esp_err_t handle_hide(
 {
     (void)node;
     (void)context;
+    (void)invocation;
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    if (params->child != NULL) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.hide accepts only {}");
     }
     cJSON_Delete(params);
     return room_canvas_hide(out_payload_json, out_error);
@@ -113,6 +155,7 @@ static esp_err_t handle_hide(
 static esp_err_t handle_snapshot(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -120,9 +163,15 @@ static esp_err_t handle_snapshot(
 {
     (void)node;
     (void)context;
+    (void)invocation;
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    static const char *const fields[] = {"format", "maxWidth", "quality"};
+    if (!only_fields(params, fields, 3)) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.snapshot contains an unknown field");
     }
 
     cJSON *format = cJSON_GetObjectItemCaseSensitive(params, "format");
@@ -157,6 +206,7 @@ static esp_err_t handle_snapshot(
 static esp_err_t handle_push_jsonl(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -167,6 +217,11 @@ static esp_err_t handle_push_jsonl(
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    static const char *const fields[] = {"jsonl"};
+    if (!only_fields(params, fields, 1)) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.a2ui.pushJSONL contains an unknown field");
     }
     cJSON *jsonl = cJSON_GetObjectItemCaseSensitive(params, "jsonl");
     if (!cJSON_IsString(jsonl) || jsonl->valuestring == NULL) {
@@ -180,6 +235,10 @@ static esp_err_t handle_push_jsonl(
         jsonl_len,
         out_payload_json,
         out_error);
+    if (err == ESP_OK) {
+        room_canvas_bind_owner_session(
+            room_canvas_surface_id != NULL ? invocation->session_key : NULL);
+    }
     cJSON_Delete(params);
     return err;
 }
@@ -187,6 +246,7 @@ static esp_err_t handle_push_jsonl(
 static esp_err_t handle_push(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -197,6 +257,11 @@ static esp_err_t handle_push(
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
+    }
+    static const char *const fields[] = {"messages", "jsonl"};
+    if (!only_fields(params, fields, 2)) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.a2ui.push contains an unknown field");
     }
 
     cJSON *messages = cJSON_GetObjectItemCaseSensitive(params, "messages");
@@ -219,6 +284,10 @@ static esp_err_t handle_push(
             out_error,
             "canvas.a2ui.push requires messages or jsonl");
     }
+    if (err == ESP_OK) {
+        room_canvas_bind_owner_session(
+            room_canvas_surface_id != NULL ? invocation->session_key : NULL);
+    }
     cJSON_Delete(params);
     return err;
 }
@@ -226,6 +295,7 @@ static esp_err_t handle_push(
 static esp_err_t handle_reset(
     esp_openclaw_node_handle_t node,
     void *context,
+    const esp_openclaw_node_command_invocation_t *invocation,
     const char *params_json,
     size_t params_len,
     char **out_payload_json,
@@ -233,17 +303,24 @@ static esp_err_t handle_reset(
 {
     (void)node;
     (void)context;
+    (void)invocation;
     cJSON *params = parse_params(params_json, params_len, out_error);
     if (params == NULL) {
         return ESP_ERR_INVALID_ARG;
     }
+    if (params->child != NULL) {
+        cJSON_Delete(params);
+        return invalid_params(out_error, "canvas.a2ui.reset accepts only {}");
+    }
     cJSON_Delete(params);
-    return room_canvas_a2ui_reset(out_payload_json, out_error);
+    esp_err_t err = room_canvas_a2ui_reset(out_payload_json, out_error);
+    if (err == ESP_OK) room_canvas_bind_owner_session(NULL);
+    return err;
 }
 
 esp_err_t room_canvas_register_node_commands(esp_openclaw_node_handle_t node)
 {
-    static const esp_openclaw_node_command_t COMMANDS[] = {
+    static const esp_openclaw_node_command_v2_t COMMANDS[] = {
         {.name = "canvas.present", .handler = handle_present},
         {.name = "canvas.navigate", .handler = handle_navigate},
         {.name = "canvas.hide", .handler = handle_hide},
@@ -254,7 +331,7 @@ esp_err_t room_canvas_register_node_commands(esp_openclaw_node_handle_t node)
     };
     for (size_t i = 0; i < sizeof(COMMANDS) / sizeof(COMMANDS[0]); ++i) {
         ESP_RETURN_ON_ERROR(
-            esp_openclaw_node_register_command(node, &COMMANDS[i]),
+            esp_openclaw_node_register_command_v2(node, &COMMANDS[i]),
             TAG,
             "register %s",
             COMMANDS[i].name);
