+++ components/esp-openclaw-room-node/room_canvas_node_cmd.h
