+++ components/esp-openclaw-room-node/room_device_commands.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "esp_openclaw_node.h"
+
+esp_err_t room_device_register_node_commands(esp_openclaw_node_handle_t node);
