+++ examples/waveshare-esp32-s3-touch-amoled-2.06-room-node/CMakeLists.txt
@@ -2,7 +2,9 @@ cmake_minimum_required(VERSION 3.16)
 
 set(EXTRA_COMPONENT_DIRS
     "../../components/esp-openclaw-node"
-    "../../components/esp-openclaw-talk")
+    "../../components/esp-openclaw-node-provisioning"
+    "../../components/esp-openclaw-talk"
+    "../../components/esp-openclaw-room-node")
 set(COMPONENTS main)
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
 project(openclaw_waveshare_room_node)
