--- examples/waveshare-esp32-s3-touch-amoled-2.06-room-node/main/Kconfig.projbuild
@@ -1,48 +1,0 @@
-menu "OpenClaw room node"
-
-config OPENCLAW_ROOM_WIFI_SSID
-    string "Wi-Fi SSID"
-    default ""
-
-config OPENCLAW_ROOM_WIFI_PASSWORD
-    string "Wi-Fi password"
-    default ""
-
-config OPENCLAW_ROOM_SETUP_CODE
-    string "Voice-node setup code"
-    default ""
-    help
-        Generate with `openclaw qr --voice-node --setup-code-only`. It is used
-        only until the node and operator device tokens have been persisted.
-
-config OPENCLAW_ROOM_GATEWAY_HTTP_BASE_URL
-    string "Gateway HTTPS origin for relative Talk offers"
-    default ""
-    help
-        HTTPS origin corresponding to the setup code WebSocket endpoint.
-        GPT-Live uses this to resolve its Gateway-relative SDP offer URL.
-        Leave empty when the configured Talk provider returns an absolute offer
-        URL. Plain HTTP exposes the one-use Talk credential and SDP on the LAN;
-        use it only for an explicitly trusted, isolated development network.
-
-config OPENCLAW_ROOM_TALK_PROVIDER
-    string "Optional Talk provider override"
-    default ""
-
-config OPENCLAW_ROOM_TALK_MODEL
-    string "Optional Talk model override"
-    default ""
-    help
-        Leave empty to use Gateway Talk configuration. GPT-Live examples are
-        gpt-live-1-codex and gpt-live-1-boulder-alpha.
-
-config OPENCLAW_ROOM_TALK_VOICE
-    string "Optional Talk voice override"
-    default ""
-
-config OPENCLAW_ROOM_CALL_IDLE_SECONDS
-    int "Maximum call lifetime in seconds"
-    range 15 1800
-    default 180
-
-endmenu
